Move test files to standardised location
