--- circle.yml
@@ -1,8 +1,0 @@
-machine:
-  php:
-    version: 7.1.6
-
-test:
-  override:
-    - mkdir -p $CIRCLE_TEST_REPORTS/phpunit
-    - vendor/bin/phpunit --log-junit $CIRCLE_TEST_REPORTS/phpunit/junit.xml
