--- phpunit.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<phpunit colors="true" bootstrap="test/unit/Helper.php">
-	<testsuites>
-		<testsuite>
-			<directory suffix=".test.php">test/unit</directory>
-		</testsuite>
-	</testsuites>
-
-	<logging>
-		<log type="coverage-text" target="php://stdout"
-			showUncoveredFiles="true" />
-		<log type="coverage-html" target="test/unit/report"
-            charset="UTF-8" highlight="false"
-            lowUpperBound="35" highLowerBound="70" />
-	</logging>
-
-	<filter>
-		<whitelist>
-			<directory suffix=".php">src</directory>
-		</whitelist>
-	</filter>
-</phpunit>
