+++ test/unit/Connection/DefaultSettingsTest.php
