+++ test/unit/Connection/SettingsTest.php
