+++ test/unit/IntegrationTest.php
@@ -4,7 +4,7 @@ namespace Gt\Database;
 use Exception;
 use Gt\Database\Connection\Driver;
 use Gt\Database\Connection\Settings;
-use Gt\Database\Test\Helper;
+use Gt\Database\Test\Helper\Helper;
 use PHPUnit\Framework\TestCase;
 
 class IntegrationTest extends TestCase {
