+++ test/unit/Query/QueryCollectionTest.php
