+++ test/unit/Result/RowTest.php
