+++ test/unit/phpunit.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<phpunit colors="true">
+	<testsuites>
+		<testsuite>
+			<directory suffix="Test.php">.</directory>
+		</testsuite>
+	</testsuites>
+
+	<logging>
+		<log type="coverage-text" target="php://stdout" showUncoveredFiles="true" />
+		<log type="coverage-html" target="./_coverage" charset="UTF-8" highlight="false" lowUpperBound="35" highLowerBound="70" />
+	</logging>
+
+	<filter>
+		<whitelist>
+			<directory suffix=".php">../../src</directory>
+		</whitelist>
+	</filter>
+</phpunit>
